topology: move SSSDTopologyMark to config
In order to avoid circular dependency in later patches.
